move api schemas to a new file
