+++ shared/schemas/api.ts
@@ -1,0 +1,66 @@
+import { z, ZodSchema } from "zod";
+
+export type OperationTag = "configs";
+
+export type EndpointMetadata = {
+  /** Authentication options, by default a bearer token is required. */
+  authenticationOptions?: RequestAuthenticationOptions;
+  openApiTags?: OperationTag | OperationTag[];
+};
+
+export type RequestAuthenticationOptions = {
+  /** Endpoint is accessible without any authentication. If `false` bearer authentication is required. */
+  isPublic?: boolean;
+  /** Endpoint is accessible with ape key authentication in  _addition_ to the bearer authentication. */
+  acceptApeKeys?: boolean;
+  /** Endpoint requires an authentication token which is younger than one minute.  */
+  requireFreshToken?: boolean;
+  noCache?: boolean;
+};
+
+export const MonkeyResponseSchema = z.object({
+  message: z.string(),
+});
+export type MonkeyResponseType = z.infer<typeof MonkeyResponseSchema>;
+
+export const MonkeyValidationErrorSchema = MonkeyResponseSchema.extend({
+  validationErrors: z.array(z.string()).nonempty(),
+});
+export type MonkeyValidationError = z.infer<typeof MonkeyValidationErrorSchema>;
+
+export const MonkeyClientError = MonkeyResponseSchema;
+export const MonkeyServerError = MonkeyClientError.extend({
+  errorId: z.string(),
+  uid: z.string().optional(),
+});
+export type MonkeyServerErrorType = z.infer<typeof MonkeyServerError>;
+
+export function responseWithNullableData<T extends ZodSchema>(
+  dataSchema: T
+): z.ZodObject<
+  z.objectUtil.extendShape<
+    typeof MonkeyResponseSchema.shape,
+    {
+      data: z.ZodNullable<T>;
+    }
+  >
+> {
+  return MonkeyResponseSchema.extend({
+    data: dataSchema.nullable(),
+  });
+}
+
+export function responseWithData<T extends ZodSchema>(
+  dataSchema: T
+): z.ZodObject<
+  z.objectUtil.extendShape<
+    typeof MonkeyResponseSchema.shape,
+    {
+      data: T;
+    }
+  >
+> {
+  return MonkeyResponseSchema.extend({
+    data: dataSchema,
+  });
+}
